Refactor pkgs/ts/circles package
