+++ pkgs/ts/circles/src/components/DialogCard.tsx
@@ -1,0 +1,45 @@
+import React, { ReactElement } from 'react';
+import tw from 'twin.macro';
+
+const Frame = tw.div`bg-gray-50 border border-pink-500 border-dotted`;
+const CardHead = tw.div`max-w-7xl mx-auto pt-12 pb-8 px-4 sm:px-6 lg:pt-16 lg:pb-4 lg:px-8 lg:flex lg:justify-between`;
+const CardBody = tw.div`max-w-7xl mx-auto pb-12 px-4 sm:px-6 lg:pb-16 lg:px-8`;
+const IntroContent = tw.div`lg:w-1/2`;
+const MainContent = tw.div`w-full`;
+const Control = tw.div`mt-8 flex lg:mt-0 lg:flex-shrink-0 lg:mt-8`;
+const Debug = tw.div`p-8`;
+
+type DemoCardProps = {
+  text: ReactElement;
+  interaction?: ReactElement;
+  control: ReactElement;
+  mainContent?: ReactElement;
+  debug?: ReactElement;
+};
+
+export const DialogCard = ({
+  text,
+  interaction,
+  control,
+  mainContent,
+  debug,
+}: DemoCardProps): ReactElement => {
+  return (
+    <>
+      <Frame>
+        <CardHead>
+          <IntroContent>
+            {text}
+            {interaction}
+          </IntroContent>
+          <Control>{control}</Control>
+        </CardHead>
+
+        <CardBody>
+          <MainContent>{mainContent}</MainContent>
+        </CardBody>
+      </Frame>
+      <Debug>{debug}</Debug>
+    </>
+  );
+};
