--- pkgs/ts/circles/src/onboarding/directionToOrientation.ts
@@ -1,11 +1,0 @@
-import { Direction } from "generated/output/CirclesPink.Garden.StateMachine.Direction";
-import { Orientation } from "../components/animation/FadeIn";
-
-export const directionToOrientation = (direction: Direction): Orientation => {
-    switch (direction.type) {
-      case 'forwards':
-        return 'left';
-      case 'backwards':
-        return 'right';
-    }
-  };
