+++ pkgs/ts/circles/src/onboarding/utils/getCounter.ts
