+++ pkgs/ts/circles/src/onboarding/utils/mapIndicatorColors.ts
