+++ pkgs/ts/circles/src/onboarding/views/index.tsx
@@ -1,0 +1,5 @@
+export { InfoGeneral } from './InfoGeneral';
+export { AskUsername } from './AskUsername';
+export { AskEmail } from './AskEmail';
+export { InfoSecurity } from './InfoSecurity';
+export { MagicWords } from './MagicWords';
