+++ pkgs/ts/storybook/src/demo/Onboarding.stories.tsx
@@ -1,12 +1,10 @@
-import React, { ReactElement } from 'react';
-import { ComponentMeta } from '@storybook/react';
-import { Onboarding } from 'circles/src/onboarding/Onboarding';
-
+import React, { ReactElement } from "react";
+import { ComponentMeta } from "@storybook/react";
+import { Onboarding } from "circles/src/onboarding";
 
 export default {
-  title: 'Components/Onboarding',
+  title: "Components/Onboarding",
   component: Onboarding,
 } as ComponentMeta<typeof Onboarding>;
-
 
 export const GardenAPI = (): ReactElement => <Onboarding />;
